refactor!: move code into a single extractor360 package (A1)
BREAKING CHANGE (imports only, CLI unchanged): the flat top-level packages
core/, ui/, utils/ and the main module are now subpackages of a single
`extractor360` package. Installing the wheel previously dropped the generic
names `core`, `ui`, `utils` and `main` into site-packages, which is asking
for collisions with any other package. Prerequisite for PyPI publication.

- pyproject: packages.find -> extractor360*, entry point
  extractor360.main:main, dynamic version attr updated
- src/main.py kept as a dev-only launcher shim so the documented
  `python3 src/main.py` invocation still works from a checkout;
  `python -m extractor360` also supported
- all imports rewritten; regression-guard test path updated
- ARCHITECTURE.md: directory tree fixed (was duplicated/corrupted) and
  updated to the new layout

Verified: 62 tests + 5 subtests pass, ruff clean, wheel contains only
extractor360/* (stale egg-info/build purged), GUI imports offscreen,
CLI e2e on synthetic 360 video unchanged (22 files incl. manifest).
