+++ src/extractor360/__init__.py
@@ -1,0 +1,2 @@
+"""360 Extractor Pro — 360°/standard media preprocessing for photogrammetry and Gaussian Splatting."""
+from extractor360.core.version import VERSION as __version__  # noqa: F401
