+++ src/extractor360/__main__.py
@@ -1,0 +1,5 @@
+"""Allow running the application as `python -m extractor360`."""
+from extractor360.main import main
+
+if __name__ == "__main__":
+    main()
