+++ src/extractor360/core/ai_classes.py
