+++ src/extractor360/core/geometry.py
