+++ src/extractor360/ui/collapsible_section.py
