+++ src/extractor360/ui/log_panel.py
