+++ src/extractor360/ui/styles.qss
