+++ src/extractor360/ui/toggle_switch.py
