+++ src/extractor360/ui/widgets.py
@@ -1,6 +1,6 @@
 from PySide6.QtWidgets import QLabel, QFrame, QVBoxLayout
 from PySide6.QtCore import Qt, Signal
-from ui.icons import get_icon
+from extractor360.ui.icons import get_icon
 
 class DropZone(QFrame):
     """
